refactor: move htmlfind to an internal directory
