+++ cmd/templ/generatecmd/testwatch/generate_test.go
@@ -20,7 +20,7 @@ import (
 
 	"github.com/a-h/templ/cmd/templ/generatecmd"
 	"github.com/a-h/templ/cmd/templ/generatecmd/modcheck"
-	"github.com/a-h/templ/htmlfind"
+	"github.com/a-h/templ/internal/htmlfind"
 	"golang.org/x/net/html"
 )
