+++ internal/htmlfind/htmlfind_test.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/a-h/templ/htmlfind"
+	"github.com/a-h/templ/internal/htmlfind"
 	"golang.org/x/net/html"
 )
